# client/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenApi specification - Portal Microservice",
        "description": "OpenApi documentation for Portal Microservice",
        "contact": {
            "name": "Miguel Cardoso",
            "url": "https://www.linkedin.com/in/miguelcardoso19/",
            "email": "miguel40cardoso@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local env"
        }
    ],
...
    "paths": {
        "/api/v1/user-credentials/update": {
            "put": {
                "tags": [
                    "User Credentials Controller"
                ],
                "summary": "Update user credentials",
                "description": "Updates the user credentials and returns the updated credentials.",
                "operationId": "update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentialsDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid credentials update data provided",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCredentialsDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access to update credentials",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCredentialsDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "User credentials updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCredentialsDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user-credentials/reset-password": {
            "post": {
                "tags": [
                    "User Credentials Controller"
                ],
                "summary": "Reset the user's password using the provided reset token",
                "description": "This method resets the user's password after validating the reset token and setting the new password.",
                "operationId": "resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "400": {
                        "description": "Invalid password format or request data"
                    },
                    "404": {
                        "description": "Invalid or expired token"
                    }
                }
            }
        },
        "/api/v1/user-credentials/register": {
            "post": {
                "tags": [
                    "User Credentials Controller"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user with provided credentials and returns an authentication response containing a JWT token.",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentialsDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid user credentials provided",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "User registered successfully, returns authentication response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user-credentials/recover-password": {
            "post": {
                "tags": [
                    "User Credentials Controller"
                ],
                "summary": "Recover password by sending a reset link to the logged user's email",
                "description": "Sends a password recovery email to the user with a reset token.",
                "operationId": "recoverPassword",
                "responses": {
                    "200": {
                        "description": "Password recovery email sent successfully"
                    },
                    "404": {
                        "description": "Email not found in the system"
                    }
                }
            }
        },
...
        "/api/v1/auth/sign-out": {
            "post": {
                "tags": [
                    "Authentication Controller"
                ],
                "summary": "Sign out and invalidate the user",
                "description": "Invalidates the current user by communicating with the Authentication microservice.",
                "operationId": "signOut",
                "responses": {
                    "204": {
                        "description": "Sign-out successful, token invalidated"
                    },
                    "401": {
                        "description": "Unauthorized - Token is missing or invalid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/sign-in": {
            "post": {
                "tags": [
                    "Authentication Controller"
                ],
                "summary": "Sign in using user credentials",
                "description": "Authenticate a user through the proxy and return a JWT token if successful.",
                "operationId": "signIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful, returns JWT token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user credentials",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh-token": {
            "post": {
                "tags": [
                    "Authentication Controller"
                ],
                "summary": "Refresh JWT token",
                "description": "Refresh the JWT token using the Authorization header through the proxy.",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired refresh token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponseDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user-credentials/delete": {
            "delete": {
                "tags": [
                    "User Credentials Controller"
                ],
                "summary": "Delete user credentials",
                "description": "Deletes the provided user credentials from the system.",
                "operationId": "delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "User credentials deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized access to delete credentials"
                    },
                    "400": {
                        "description": "Invalid credentials data for deletion"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserCredentialsDTO": {
                "required": [
                    "dateOfBirth",
                    "email",
                    "firstName",
                    "gender",
                    "lastName",
                    "nationality",
                    "nif",
                    "password",
                    "phoneNumber",
                    "roles"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique identifier (UUID) for each entity instance",
                        "format": "uuid",
                        "example": "e12a567d-32f8-4e9a-9073-6781f9d5e423"
                    },
                    "firstName": {
                        "type": "string",
                        "description": "First name of the user",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "description": "Last name of the user",
                        "example": "Keen"
                    },
                    "nif": {
                        "type": "string",
                        "description": "Tax Identification Number (NIF) of the user",
                        "example": "123456789"
                    },
                    "email": {
                        "pattern": "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$",
                        "type": "string",
                        "description": "Email address of the user",
                        "example": "john.keen@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "Password of the user",
                        "example": "P@ssw0rd123"
                    },
                    "newPassword": {
                        "type": "string",
                        "description": "New password for updating credentials (if applicable)",
                        "example": "NewP@ssw0rd123"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "description": "Date of birth of the user in YYYY-MM-DD format",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "nationality": {
                        "type": "string",
                        "description": "Nationality of the user",
                        "example": "Portuguese",
                        "enum": [
                            "Afghan",
                            "Albanian",
                            "Algerian",
                            "Andorran",
                            "Angolan",
                            "Argentine",
                            "Armenian",
                            "Australian",
                            "Austrian",
                            "Azerbaijani",
                            "Bahraini",
                            "Bangladeshi",
                            "Belarusian",
                            "Belgian",
                            "Belizean",
                            "Beninese",
                            "Bhutanese",
                            "Bolivian",
                            "Bosnian",
                            "Botswanan",
                            "Brazilian",
                            "Bruneian",
                            "Bulgarian",
                            "Burkinabe",
                            "Burundian",
                            "Cambodian",
                            "Cameroonian",
                            "Canadian",
                            "Capeverdean",
                            "Centralafrican",
                            "Chadian",
                            "Chilean",
                            "Chinese",
                            "Colombian",
                            "Comorian",
                            "Congolese",
                            "CongoleseDRC",
                            "Costarican",
                            "Ivorian",
                            "Croatian",
                            "Cuban",
                            "Cypriot",
                            "Czech",
                            "Danish",
                            "Djiboutian",
                            "Dominican",
                            "DominicanRep",
                            "Ecuadorian",
                            "Egyptian",
                            "Salvadoran",
                            "Equatorialguinean",
                            "Eritrean",
                            "Estonian",
                            "Ethiopian",
                            "Fijian",
                            "Finnish",
                            "French",
                            "Gabonese",
                            "Gambian",
                            "Georgian",
                            "German",
                            "Ghanaian",
                            "Greek",
                            "Grenadian",
                            "Guatemalan",
                            "Guinean",
                            "Bissaugunean",
                            "Guyanese",
                            "Haitian",
                            "Honduran",
                            "Hungarian",
                            "Icelandic",
                            "Indian",
                            "Indonesian",
                            "Iranian",
                            "Iraqi",
                            "Irish",
                            "Israeli",
                            "Italian",
                            "Jamaican",
                            "Japanese",
                            "Jordanian",
                            "Kazakh",
                            "Kenyan",
                            "Kiribati",
                            "Korean",
                            "Kuwaiti",
                            "Kyrgyz",
                            "Lao",
                            "Latvian",
                            "Lebanese",
                            "Basotho",
                            "Liberian",
                            "Libyan",
                            "Liechtenstein",
                            "Lithuanian",
                            "Luxembourgish",
                            "Malagasy",
                            "Malawian",
                            "Malaysian",
                            "Maldivian",
                            "Malian",
                            "Maltese",
                            "Marshallese",
                            "Mauritanian",
                            "Mauritian",
                            "Mexican",
                            "Micronesian",
                            "Moldovan",
                            "Monacan",
                            "Mongolian",
                            "Montenegrin",
                            "Moroccan",
                            "Mozambican",
                            "Namibian",
                            "Nauruan",
                            "Nepali",
                            "Dutch",
                            "Newzealander",
                            "Nicaraguan",
                            "Nigerien",
                            "Nigerian",
                            "Norwegian",
                            "Omani",
                            "Pakistani",
                            "Palauan",
                            "Panamanian",
                            "Papuanewguinean",
                            "Paraguayan",
                            "Peruvian",
                            "Filipino",
                            "Polish",
                            "Portuguese",
                            "Qatari",
                            "Romanian",
                            "Russian",
                            "Rwandan",
                            "Kittitian",
                            "Saintlucian",
                            "Vincentian",
                            "Samoan",
                            "Sanmarinese",
                            "Saotomean",
                            "Saudi",
                            "Senegalese",
                            "Serbian",
                            "Seychellois",
                            "Sierraleonean",
                            "Singaporean",
                            "Slovak",
                            "Slovenian",
                            "Solomonislander",
                            "Somali",
                            "Southafrican",
                            "Spanish",
                            "Srilankan",
                            "Sudanese",
                            "Surinamese",
                            "Swazi",
                            "Swedish",
                            "Swiss",
                            "Syrian",
                            "Taiwanese",
                            "Tajik",
                            "Tanzanian",
                            "Thai",
                            "Togolese",
                            "Tongan",
                            "Trinidadian",
                            "Tunisian",
                            "Turkish",
                            "Turkmen",
                            "Tuvaluan",
                            "Ugandan",
                            "Ukrainian",
                            "Emirati",
                            "British",
                            "American",
                            "Uruguayan",
                            "Uzbek",
                            "Nivanuatu",
                            "Venezuelan",
                            "Vietnamese",
                            "Yemeni",
                            "Zambian",
                            "Zimbabwean"
                        ]
                    },
                    "gender": {
                        "type": "string",
                        "description": "Gender of the user",
                        "example": "Male",
                        "enum": [
                            "Male",
                            "Female"
                        ]
                    },
                    "phoneNumber": {
                        "type": "string",
                        "description": "Phone number of the user",
                        "example": "912954239"
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "description": "Roles assigned to the user",
                        "example": [
                            "ADMIN",
                            "EMPLOYEE"
                        ],
                        "items": {
                            "type": "string",
                            "description": "Roles assigned to the user",
                            "example": "[\"ADMIN\",\"EMPLOYEE\"]"
                        }
                    }
                },
                "description": "User credentials for registration"
            },
...
            "ResetPasswordRequestDTO": {
                "required": [
                    "newPassword",
                    "token"
                ],
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "The password reset token, usually sent via email for validation.",
                        "example": "abc123xyz"
                    },
                    "newPassword": {
                        "type": "string",
                        "description": "The new password to set for the user.",
                        "example": "NewSecurePassword123!"
                    }
                },
                "description": "Reset password request containing the reset token and the new password"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insert the JWT bearer token",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}